Перенест загрузку избранных фильмов пользователя в ДАТА редьюсер
