+++ src/reducer/user/user.js
@@ -1,18 +1,15 @@
 import {extend} from "../../utils.js";
 import {AuthorizationStatus} from "../../consts.js";
 import UserModel from "../../models/user.js";
-import MovieModel from "../../models/movie.js";
 
 const initialState = {
   authorizationStatus: AuthorizationStatus.NO_AUTH,
   profile: {},
-  favoriteMovies: [],
 };
 
 const ActionType = {
   SET_AUTHORIZATION: `SET_AUTHORIZATION`,
   SET_USER_PROFILE: `SET_USER_PROFILE`,
-  SET_USER_FAVORITE_MOVIES: `SET_USER_FAVORITE_MOVIES`,
 };
 
 const ActionCreator = {
@@ -28,12 +25,6 @@ const ActionCreator = {
       payload: profile
     };
   },
-  setUserFavoriteMovies: (movies) => {
-    return {
-      type: ActionType.SET_USER_FAVORITE_MOVIES,
-      payload: movies,
-    };
-  }
 };
 
 const reducer = (state = initialState, action) => {
@@ -45,10 +36,6 @@ const reducer = (state = initialState, action) => {
     case ActionType.SET_USER_PROFILE:
       return extend(state, {
         profile: action.payload,
-      });
-    case ActionType.SET_USER_FAVORITE_MOVIES:
-      return extend(state, {
-        favoriteMovies: action.payload
       });
   }
   return state;
@@ -73,13 +60,6 @@ const Operation = {
         dispatch(ActionCreator.setUserProfile(UserModel.parseUser(data)));
       });
   },
-
-  loadFavoriteMovies: () => (dispatch, getState, api) => {
-    return api.get(`/favorite`)
-      .then(({data}) => {
-        dispatch(ActionCreator.setUserFavoriteMovies(MovieModel.parseMovies(data)));
-      });
-  }
 };
